+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Add `id` field to TimeEntry model.
+- Add `stopTimer` method to stop a running timer.
+- Add `getRunningTimer` method to retrieve the currently running timer.
+
 ## 1.0.1
 
 - Allow TimeInterval to have null end dates.